refactor(model-server): move HTML handlers explorer to own package
To avoid confusing API handlers with HTML request handlers in the
future, move the content explorer HTML handlers to their own package.
