+++ model-server/src/main/kotlin/org/modelix/model/server/Main.kt
@@ -60,17 +60,17 @@ import org.modelix.authorization.NoPermissionException
 import org.modelix.authorization.NotLoggedInException
 import org.modelix.authorization.installAuthentication
 import org.modelix.model.InMemoryModels
-import org.modelix.model.server.handlers.ContentExplorer
 import org.modelix.model.server.handlers.DeprecatedLightModelServer
 import org.modelix.model.server.handlers.HealthApiImpl
-import org.modelix.model.server.handlers.HistoryHandler
 import org.modelix.model.server.handlers.HttpException
 import org.modelix.model.server.handlers.IdsApiImpl
 import org.modelix.model.server.handlers.KeyValueLikeModelServer
 import org.modelix.model.server.handlers.MetricsApiImpl
 import org.modelix.model.server.handlers.ModelReplicationServer
 import org.modelix.model.server.handlers.RepositoriesManager
-import org.modelix.model.server.handlers.RepositoryOverview
+import org.modelix.model.server.handlers.ui.ContentExplorer
+import org.modelix.model.server.handlers.ui.HistoryHandler
+import org.modelix.model.server.handlers.ui.RepositoryOverview
 import org.modelix.model.server.store.IgniteStoreClient
 import org.modelix.model.server.store.InMemoryStoreClient
 import org.modelix.model.server.store.IsolatingStore
